immigration-bot/requirements.txt
fastapi
uvicorn
python-dotenv
openai
reductoai
requests
python-multipart
pymongo[srv]
tqdm
voyageai
umap-learn
numpy
